pyproject.toml
[tool.poetry]
name = "matcha"
version = "0.1.0"
description = ""
authors = ["daiti0113 <omochi.ds.0113@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.35.3"
langchain = "^0.2.5"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
